ArtifactUtils

ArtifactUtils.add_artifact!

ArtifactUtils.artifact_from_directory

ArtifactUtils.communicate

ArtifactUtils.copy_mode

ArtifactUtils.gist_from_file

ArtifactUtils.git_empty_history

ArtifactUtils.upload_to_gist

ArtifactUtils.with_new_gist

ArtifactUtils.add_artifact! — Method

function add_artifact!(
    artifacts_toml::String, name::String, tarball_url::String;
    clear=true,
    platform::Union{Platform,Nothing}=nothing,
    lazy::Bool=false,
    force::Bool=false
)

Downloads tarball from tarball_url, extracts it and adds it as an artifact with name name to the file artifacts_toml. If clear is true, the artifact itself is deleted afterwards. The rest of the keyword arguments are passed to Pkg.Artifacts.bind_artifact!.

From its docstring:

If platform is not nothing, this artifact is marked as platform-specific, and will be a multi-mapping. It is valid to bind multiple artifacts with the same name, but different platforms and hash'es within the same artifacts_toml. If force is set to true, this will overwrite a pre-existant mapping, otherwise an error is raised.

[...] If lazy is set to true, even if download information is available, this artifact will not be downloaded until it is accessed via the artifact"name" syntax, or ensure_artifact_installed() is called upon it.

ArtifactUtils.artifact_from_directory — Method

artifact_from_directory(source; follow_symlinks = false) -> artifact_id::SHA1

Create an artifact from the source directory and return the artifact_id.

ArtifactUtils.communicate — Method

communicate(cmd, input) -> output::Vector{UInt8}

Run cmd, write input to its stdin, and return the bytes output read from its stdout. stderr is redirected to Base.stdout.

ArtifactUtils.copy_mode — Method

copy_mode(src, dst)

Copy mode of src to dst recursively in a way compatible to how Pkg (Git) compute the tree hash.

See Pkg.GitTools.gitmode comment for how the file mode is handled specially in Windows: https://github.com/JuliaLang/Pkg.jl/blob/247a4062bfde19d93bdcbaccc9737df496fd0c2b/src/GitTools.jl#L189-L192

copy_mode is based on: https://github.com/JuliaIO/Tar.jl/blob/6a946029685639b69ce5a7cc4c4a6c0e6c6b2697/src/extract.jl#L145-L153

ArtifactUtils.gist_from_file — Method

gist_from_file(filepath::AbstractString; private::Bool = true) -> fileurl::String

Create a new gist from file at filepath. Return the "raw" HTTPS URL of the file.

ArtifactUtils.git_empty_history — Method

git_empty_history(git_dir)

Empty the history of the current branch. Create an empty commit to start fresh.

ArtifactUtils.upload_to_gist — Function

upload_to_gist(
    artifact_id::SHA1,
    [tarball];
    private::Bool = true,
    honor_overrides = false,
    # Following options are aviailable only when `tarball` is not specified:
    name::AbstractString = "$artifact_id.tar.gz",
    extension::AbstractString = ".tar.gz",
) -> gist

Create an artifact archive at path tarball (or in a temporary location) and upload it to gist. The returned value gist can be passed to add_artifact!.

Extended help

Examples

using ArtifactUtils
add_artifact!("Artifact.toml", "name", upload_to_gist(artifact_from_directory("source")))

creates an artifact from files in the "source" directory, upload it to gist, and then add it to "Artifact.toml" with the name "name".

Keyword Arguments

private: if true, upload the archive to a private gist

name: name of the archive file, including file extension

extension: file extension of the tarball. It can be used for specifying the compression method.

honor_overrides: see Pkg.Artifacts.archive_artifact

ArtifactUtils.with_new_gist — Method

with_new_gist(f; private::Bool = true) -> repo_http::AbstractString

Create a new gist, check it out as a local git repository at a temporary directory git_dir, call f(git_dir), and then push it to remote. Return the canonical gist HTTPS URL.